fix: converge user creation paths to ensure authorization is applied
All three user creation paths (GraphQL self-registration, GraphQL admin,
REST identity-resolve) now follow the same finalization logic via
`finalizeUserRegistration()` in RegistrationService.

This fixes a bug where users created via the REST identity-resolve
endpoint were missing authorization policies (empty credentialRules),
causing "AuthorizationPolicy without credential rules" errors.

Changes:
- Move authorization logic from resolver to RegistrationService
- Add `finalizeUserRegistration()` method that handles:
  - Granting essential credentials (GLOBAL_REGISTERED)
  - Applying user authorization policy
  - Applying account authorization policy
  - Processing pending invitations
  - Sending user created notification
- Remove duplicate `processCreatedUser()` and `userCreatedEvents()` from resolver
